Template method pattern to ease further refactorings.
